Add calculate-weights.R and outputted weight csv's
Add calculate-weights.R script, which takes scores/scores.csv and uses it to
generate the following weights files:
- weights/constant-weights.csv
- weights/target-and-region-based-weights.csv
- weights/target-based-weights.csv
- weights/target-type-based-weights.csv

It currently assigns -Inf scores to missing forecasts.

This script will need to be updated to use only component models and the
standardized model ID's.

If scores are added from a partially observed season, it will need an update to
use scores only from the desired training seasons, and to make sure that the
prospective forecast season is assigned the correct label.