Use issue week 28 or nearby data as ground truth rather than current
Fix ground truth target calculations to use data from issue week 28; or close to
it for seasons before 2013 when backfill records may not be fully available.

For seasons 2013/2014 on: in-season data are always from issue week 28.

For seasons 2010/2011--2013/2014: in-season data end up being either from issue
week 28 or issue week 20, as issue weeks 21..28 were not recorded in at least
the HHS Regions. (Another way would be to try to look ahead and grab data from
the closest issue week >= 28 (probably around the following issue week 40)
instead; this is not implemented, though.)